--- src/Components/Icon/index.ts
@@ -1,5 +1,0 @@
-import Cart from "./Cart";
-
-const Icon = { Cart };
-
-export default Icon;
