+++ src/Components/index.ts
@@ -1,9 +1,9 @@
 import { Animation, Button, CheckBox, Confirm, Modal, PageTitle, SubmitBox } from "./@shared";
 
 import { default as CartAnimation } from "./CartAnimation";
+import { default as CartIcon } from "./CartIcon";
 import { default as CartItem } from "./CartItem";
 import { default as Header } from "./Header";
-import { default as Icon } from "./Icon";
 import { default as OrderList } from "./OrderList";
 import { default as OrderProductItem } from "./OrderProductItem";
 import { default as Portal } from "./Portal";
@@ -11,4 +11,4 @@ import { default as Product } from "./Product";
 import { default as ProductImage } from "./ProductImage";
 
 export { Animation, Button, CheckBox, Confirm, Modal, PageTitle, SubmitBox };
-export { CartAnimation, CartItem, Header, Icon, OrderList, OrderProductItem, Portal, Product, ProductImage };
+export { CartAnimation, CartItem, Header, CartIcon, OrderList, OrderProductItem, Portal, Product, ProductImage };
